+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 
 ### 🐛 Bug Fixes
 
-- Look up sources for partially verified smart contracts [#12221](https://github.com/blockscout/blockscout/pull/12221)
+- Look up sources for partially verified smart contracts ([#12221](https://github.com/blockscout/blockscout/pull/12221))
 - BufferedTask-based approach for fetching Arbitrum-specific settlement info ([#12192](https://github.com/blockscout/blockscout/pull/12192))
 - Contract creation transaction associations and bytecode twin detection ([#12086](https://github.com/blockscout/blockscout/issues/12086))
 - Improve background migrations + new `Indexer.Migrator.RecoveryWETHTokenTransfers` ([#12065](https://github.com/blockscout/blockscout/issues/12065))